move transform view to common
