+++ include/dr/mhp.hpp
@@ -42,7 +42,9 @@ namespace rng = ranges;
 #include "concepts/concepts.hpp"
 #include "details/view_detectors.hpp"
 #include "details/segments_tools.hpp"
-#include "details/views.hpp"
+#include "views/views.hpp"
+#include "details/normal_distributed_iterator.hpp"
+#include "views/transform.hpp"
 #include "details/memory.hpp"
 #include "details/halo.hpp"
 #include "mhp/cpos.hpp"
