+++ include/dr/shp/algorithms/execution_policy.hpp
@@ -5,7 +5,6 @@
 #pragma once
 
 #include <CL/sycl.hpp>
-#include <dr/details/ranges_shim.hpp>
 #include <span>
 #include <vector>
