+++ include/dr/shp/shp.hpp
@@ -4,8 +4,10 @@
 
 #pragma once
 
-#include <dr/details/ranges_shim.hpp>
-
+#include "../details/normal_distributed_iterator.hpp"
+#include "../details/ranges_shim.hpp"
+#include "../details/segments_tools.hpp"
+#include "../views/transform.hpp"
 #include "algorithms/execution_policy.hpp"
 #include "algorithms/for_each.hpp"
 #include "algorithms/reduce.hpp"
