+++ include/dr/shp/views/standard_views.hpp
@@ -7,8 +7,8 @@
 #include <dr/details/segments_tools.hpp>
 #include <dr/shp/distributed_span.hpp>
 #include <dr/shp/views/enumerate.hpp>
-#include <dr/shp/views/transform.hpp>
 #include <dr/shp/zip_view.hpp>
+#include <dr/views/transform.hpp>
 
 namespace shp {
