+++ include/dr/shp/zip_view.hpp
@@ -6,10 +6,8 @@
 
 #include <oneapi/dpl/iterator>
 
-#include <dr/concepts/concepts.hpp>
 #include <dr/details/iterator_adaptor.hpp>
 #include <dr/details/owning_view.hpp>
-#include <dr/details/ranges_shim.hpp>
 #include <dr/details/view_detectors.hpp>
 #include <ranges>
